Add matplotlib-base to dependencies
matplotlib has been added to the environment.yml file, so I'm adding
matplotlib-base to the dependencies in meta.yaml. I'm not adding
the full matplotlib package because it also pulls in the pyqt stack,
which I don't think we need. I will also open an issue in the main
repo to clarify the matplotlib vs. matplotlib-base question.